[flow] Kill TypeT subtyping rules
Summary:
Now that we have a dedicated namespace type that ensures by construction that types don't leak into value space, and we have a dedicated `ValueToTypeReferenceT` constraint to do the transform, we have successfully eliminated the leak of `TypeT` to the external system (they shouldn't get past `mk_nominal_type` after we have done the transform). Hence, it's the time to kill all those subtyping rules around `TypeT`, since they never make sense, and eliminating them can help us to avoid accidentally depending on them in the future.

Can we go one more step to kill `TypeT` completely? I think we can eventually, but I have found that some normalization code is still heavily depending on `TypeT` and various `type_t_kind`. While they can be replaced with reasons, we know reasons are unreliable, so we should probably stabilize reason first before doing that. In addition, the current state already limits the damage of `TypeT` to a very small part of the codebase, so it's a good time to stop.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54100994

fbshipit-source-id: 4564d83cf094ee7c35add45a9b4ce376e6ef4171