[flow][refactor] Organize various environment during type annotation generation into a single object
Summary:
We seem to have a proliferation of environment we like to push down when we are generating types for type annotations. I plan to soon add another one related to `NoInfer` intrinsic. Therefore, I plan to first simplify the code to make it less painful.

In this diff, I organized tparams, infer_tparams and in_renders_arg into an env object. It technically results in some subtle behavior change, since in this implementation `in_renders_arg` will be pushed down more comprehensively. I find the current behavior of only pushing it down to certain type of annotations quite arbitrary (e.g. for unions but not for maybe).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54164848

fbshipit-source-id: 64a626813c725dae4fc288b610d7ea7f0791faa0