[flow] make typed AST in ty_normalizer_imports optional
Summary:
The Ty_normalizer uses the typed AST to recover information about imported values and types to determine if the various symbols it encounters are imported as some symbol or not (Remote). While it is convenient to use the TAST this way when it is available, we don't really need that entire structure to get this information.

This diff makes the `typed_ast` parameter of Ty_normalizer_imports optional, and uses other ways to recover this information.

Why would we not want to pass the typed AST?

Commands like autocomplete currently compute the typed AST over a slightly changed input file to produce their result. This means that the cached typed AST cannot be reused and a new one needs to be computed. So, if we can avoid the need for the typed AST to process the command in the first place, we can potentially speed up processing times. (Reminder: building the TAST requires a full check of the AST.) The type normalizer, that this diff is targeting, is one of the dependencies of autocomplete code on the typed AST.

When typed_ast is not available in Ty_normalizer_imports, the necessary information is recovered from Import_export.

The results should be identical with the exception of deeply destructured requires
```
const { f: { g } } = require("m");
```
This is possible to implement in the new code, but it would be a bit more involved and pretty unlikely to happen in practice.

I'm keeping both versions available, since looking up information in the typed AST, when available, should still be faster.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54332603

fbshipit-source-id: 07f3650737fa63cecdaabee5180652a5c85f79d1