[flow][autocomplete] use type_env instead of TAST in local_value_identifiers
Summary:
Autocomplete looks up local variables in scope to include in its result. It does so by looking up the current enclosing scope for locations and names and then looks up the types associated with those locations in the typed AST. Instead of performing a full pass over the typed AST we can instead look up the names in the typed environment. This change also helps move away from the dependency of autocomplete code from the typed AST.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D53746831

fbshipit-source-id: 86106ae2676882cbc620874b984b8332c181a747