[flow] add internal --do_not_use_typed_AST_for_imports flag to type-at-pos
Summary:
The purpose of this diff is to add regression testing for D54332603. Currently, the TAST is available for all commands and so it not practical to test any command that uses normalization with it.

This diff adds a flag `--do_not_use_typed_AST_for_imports` to the type-at-pos command intended for internal testing purposes. When provided, type-at-pos will not use the typed AST when normalizing imports.

I have ported tests/type_at_pos_imports to use both the old and the new version.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54332601

fbshipit-source-id: cccbbeea64181649b3e27e0f0649825995f1b60a