[flow] Track pruned deps in context per file
Summary:
In this diff, we start to track pruned transitive dependencies that a file actually uses to type check itself. We do this by adding a file to the context when we forced it when lazily merging in type sigs.

This is used to help to prune down a list of actually used transitive deps, so that we can use it to comprehensively index all glean transitive dependencies that are actually necessary, instead of all transitive dependencies as determined by import-export alone.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54377835

fbshipit-source-id: 8003478e4bb1b7d23b6b6a065a833fc33aaed725