[flow][glean] Add support for experimental `--include-pruned-transitive-deps` flag
Summary:
This diff adds an experimental mode that will try to additional index transitive dependencies that are actually useful for checking a file.

We find such transitive dependencies with the information collected in D54377835. After the initial run, we will do a second run on the union of such transitive deps, which hopefully contains all the information a partial glean indexing needs.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54377834

fbshipit-source-id: 47e4308daab5e14c4c8cb6b8c30d1510c148435e