[flow][ide] Fix bug where we leak AUTO332 impl detail
Summary:
D52887771 accidentally breaks it. The empty tvar is there not for magically accumulating bounds, but it seems to designed to ensure we don't suggest the type of this literal we are completing, which contains AUTO332. We don't have to use empty tvar for this. A resolved EmptyT is enough.

Changelog: [fix][ide] Fixed a bug where AUTO332 appeared in the result of string literal type completion.

Reviewed By: gkz

Differential Revision: D54659383

fbshipit-source-id: af32da42ee63b800bc692e689fb0cd96669c2c52