[flow] add tests on synthetic generic behavior
Summary:
The errors in the added examples rely on information stored in `Subst_name.Synthetic`. This variant was originally added to support spread of generics. Since then more cases have been added that are not related to spreading. The error messages that are produced in those cases, however, still assume that the generics are being spread, potentially leading to confusion. This diff adds some tests to expose the issue. Later diff improves the messaging to make it not assume spreading.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D54708915

fbshipit-source-id: cd826775f261f6f1a47ddb92d413b9b71d607061