[flow] add kind to Subst_name.Synthetic
Summary:
Adds some more information to `Subst_name.t`, specifically the kind of operation that is responsible for the synthetic generic. The goal here is to use this to improve error messages and type normalization. Error messages are improved since we do not assume any more that the operations is always spreading. I'm using generic phrasing to refer to the other cases:
```
Cannot [...] is incompatible with type including generic type(s) `T` [2].
```

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54668884

fbshipit-source-id: 82639a4d74d789a1791b54b27ac83b0bb6483a46