[flow] improve formatting of Subst_name.Synthetic
Summary:
The string stored as "name" in `Subst_name.Synthetic` was initially expected to be code-like so that when rendered we'd be able to surround it in backticks and render it as code. This was not always the case, however, (e.g. see tests changed in this diff) and especially after the next diff it won't be the case even more.

This diff improves the behavior of this synthetic name by making `Generic.to_string` return an already formatted string (with backticks included). This is by no means a particularly satisfying solution. That would probably require encoding the name in a special structure that would be serialized appropriately when this name makes itself to a reason or an error. This is a bigger change and perhaps not as worthwhile at this point.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54758550

fbshipit-source-id: 4a24365e4377e5fa112c0f5986f413f48e2c980d