[flow] better normalization of GenericT Synthetic
Summary:
This diff improves the normalization of "synthetic" `GenericT`s.

Before, when we happened to need to normalize a synthetic generic, for example in the "evaluated" version of
```
type F<T: {...}> = {...T, o: O, ...}
```
we would fail to find `T` in the tparams environment (since `T` was a synthetic generic whose actual name does not match the one from the env). We would end up using the bound of the generic to proceed with normalization instead. This approach often leads to confusing results. So the evaluated version of this type used to be `{ o: O }`.

This diff tries harder to recover a descriptive type from the information in the synthetic `GenericT` (some of this was added in D54668884). So in the above example, we are able to recover the more descriptive `{...T, o: O, ...}`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54668883

fbshipit-source-id: 85241dd94448ef50c90c28b3687bf1ddd11fa555