[flow] do not use tparams_rev in normalization
Summary:
When the normalizer was first written, normalizing the type of `x` in code like
```
function foo<X: T>(x: X) {
  return x;
}
```
required knowledge of what type parameters were in scope. Back then, Flow would check `foo` by considering the two possible cases of `X` (`empty` and `T`), a process called "generate-tests". The type of `x` that was recorded on the typed AST would be `T` with a reason that hinted to the fact that `T` was a bound of the type parameter. We would then try to match it with type parameters that were in scope, so that we would end up normalizing the result back to `X`.

It was important to keep that tparam environment in place to distinguish this case from
```
const x = foo<number>(1);
```
In this case the same type we computed for `x` within `foo` would lead to the result of the function call. However, here, we would like the type for `x` to not be `X`, but instead the bound `number` (in place for `T`). Indeed, this is what we'd get, given that `X` was not in tparams env anymore.

With new generics, and especially with LTI's annotation requirement, we no longer need a tparam env to get the same result. The type of `x` with `foo` is a `GenericT`, whereas outside of `foo` it will be a concrete type.

We can therefore directly normalize `GenericT`s to `Ty.Bound`.

The only case we'd want to get something different is when the `GenericT` comes from an `infer T` annotation in the condition of a conditional type, so we're keeping `infer_tparams`.

With this change we also error more aggressively in the normalizer when we try to normalize a `GenericT` with a `Synthetic` name. The reason is that the name produced in these cases is typically not a good one to use for normalization. From the error diff in the unit tests, this tends to occur in the evaluated version of a type, specifically when we force EvalTs in abstract context (under a bound). Before, we would use the type of the bound to produce a result. I don't think this is a useful output, so instead we now error (don't return a result)

This change also lets us do some serious cleanup since we don't need to compute the `tparams_rev` for every input to the normalizer (coming up).

Changelog: [internal]

Reviewed By: jbrown215

Differential Revision: D54476871

fbshipit-source-id: 7a9d60699d567b104f14b0738e43a70ad48e5338