[flow][autocomplete] Autocomplete on string literal types returns empty result
Summary:
I tried to figure out what the intention was behind capturing the case of `StringLiteral` type in autocomplete, given that we never return an actual result.

In D18509589 and D18616795, we started returning `Ac_literal` as the `autocomplete_type`. This did not return an actual result and so we would log these as `AcEmpty`. This answered the question about why we caught this case in the first place.

In D21183466 we started returning results for *some* literals. We still did not return any meaningful result for string literal types, but this case was now also mapped to `AcReturn` result since the same `Ac_literal` constructor was still used.

Finally, with D54659383 the type returned used for auto-completing string type literals was hard-coded to `empty`.

This diff restores the original behavior of returning `AcEmpty` when auto-completing in a string literal type. It also reverts the change of D54659383, so that we don't have to use the hook in type_annotation.ml.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54884663

fbshipit-source-id: 29dfc8af6063e8e953bb7068e9cf942411f183b5