[flow] add type_parameter_mapper_generic
Summary:
`type_parameter_mapper` extends the polymorphic AST mapper with the capability of tracking a list of type parameters that are in scope at each point of the AST. Until now, these type params were of type `Type.typeparam list` which necessitated that the pass was over a typed AST. This is a bit limiting for cases where we want to get the list of type params but don't have a typed AST available. (I'm working on changes to autocomplete code that will introduce this restriction.)

This diff addresses this by abstracting `type_parameter_mapper` over the type of the parameter (this is the new `'P` parameter). The new virtual class `type_parameter_mapper` will now require the `make_typeparam` method (among a few other) is provided.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54910378

fbshipit-source-id: c9f77d4ada5e96771e88181ddf143dd9ca0fc5e7