[flow][autocomplete] only strings in tparams_rev in process_location_result
Summary:
AutocompleteService_js only uses tparams_rev for the names of the parameters. Instead of computing a list of `Type.typeparam`s and then accessing just their names, this diff changes `process_request_searcher` to only gather the names of the type parameters (as strings) from the beginning. It does this by instantiating `type_parameter_mapper_generic` with `'P` as `string`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D54910377

fbshipit-source-id: b1fb00243b2168161759812a4784e8c73fbd0bfb