[flow] No longer make libdef parsing error stop the world
Summary:
Currently, if we fail to parse libdef files (determined by failed to find them in the parsing heap), we will stop all work (no checking at all).

I managed to trace the behavior back to D2176600 when it's first introduced. It's still unclear to me why we have to do this. Maybe we want to show all the libdef errors up front rather than make it hidden in a sea of other errors, but it would be better achieved at the error printing stage.

Since the end of last year, the handling of libdef has switched to a new system: it's treated mostly as a normal file, except that we will run a special type sig merge to extract the signatures of the globals. Under the new system, the stop-the-world behavior seems even more weird and inconsistent, especially consider that we only stop the world for parse but not checking error for libdef.

Therefore, I decide to remove this behavior in this diff.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54989487

fbshipit-source-id: 3dd2977dce25d12f93399502d004b02bb87474e0