[ci] Properly test mac and linux in GitHub Actions CI (#9130)
Summary:
Currently, the mac and linux tests are passing, but if we look at the logs, they are failing for every single test. It turns out that we are not even downloading the artifacts from previous build jobs. We also need to set the executable bit to actually make the tests runnable (this is not needed in circle ci jobs and also not needed for windows builds, so it's unclear why, which needs further investigation)

Pull Request resolved: https://github.com/facebook/flow/pull/9130

Test Plan: See the CI on GitHub

Reviewed By: gkz

Differential Revision: D55020261

Pulled By: SamChou19815

fbshipit-source-id: 877195f40c22a0f408aebb72eb9f60f59ec28733