[flow] add enclosing_node_mapper
Summary:
This diff adds a new class that extends the `Flow_polymorphic_ast_mapper`. The new functionality is a way to access the closest expression or statement visited during an AST traversal.

This feature will be used in a new mode for autocomplete that does not have access to the entire typed AST. Instead it will create typed versions of AST nodes. Given that we only have entrypoints in the statement.ml for specific kinds of AST nodes (mainly expressions and statements), we need a way to compute the closest expression or statement at any point during an AST traversal.

This diff also adds a couple of helper functions that operate on the result of this searcher.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55047334

fbshipit-source-id: 78652dccfef0a0aa9ef9c5a857a6eca76f898cd9