[flow][refactor] Make `Files.options` opaque
Summary:
These files logic is the most complicated parts of Flow that's hard to understand. To make it easier to understand, we want to easily tell how are these configs actually interpreted in the code. However, with the detail of the options fully transparent, it's hard to easily find where the options are inspected.

In this diff, I make options opaque, so that all inspections have to be done through functions, which we can grep much more easier. Mostly importantly, I did find that `declarations` in file options is only used in one place, which can help future refactors.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55049081

fbshipit-source-id: 6ec1a0ca86803c306aa62a9c4ca5b0669677f315