[flow][autocomplete][refactor] jsx_element in process_request_searcher
Summary:
This diff refactors the mapper to override `jsx_element` instead of `jsx_opening_element` and `jsx_closing_element`. The reason is that in a later diff we will pass the JSX expression over to `visit_jsx_opening_element` and `visit_jsx_closing_element` so that the on-demand mapper can call inference on JSX expression directly (`(loc, Ast.Expression.JSXElement expr)`).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55053040

fbshipit-source-id: 2a6ccde253666b7445ae5012aaa017a87683dd24