[flow][autocomplete][refactor] abstract autocomplete_js over annot position
Summary:
This diff follows the same approach as D55053045, but for `Autocomplete_js.process_request_searcher`. Like the previous diff, this one also introduces a number of virtual methods for accessing locations and type annotations from the now abstract annotation position `'T`.

Because computing types will not be cheap like accessing a type on the typed AST, in a few places, we introduce lazy thunks to compute the types where convenient and then force the types if we end up needing them (e.g. when the search location matches a location on the AST).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55053039

fbshipit-source-id: 2272fe03555d373edbb5937211098e6a47bc3beb