[flow] on-demand get_def_process_location.ml
Summary:
This diff implements the on-demand version of `Get_def_process_location` searcher, by filling in the latter's virtual methods. To do so, it leverages the `enclosing node_searcher`. Specifically, the pervasive method `type_from_enclosing_node` does the following: First, it gets the closest enclosing node, which is either a statement or an expression. It then calls into statement.ml inference to computed the typed AST version of this node, and, finally, it looks up a specific target location location therein.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55053049

fbshipit-source-id: 56871d5cbbfc2dda7e3f4f960f276de2d99871fd