[flow][autocomplete] include process_request_searcher to mli
Summary:
This diff includes `process_request_searcher` to autocomplete_js.mli. It is not necessary to export this type for other modules to use, but it in upcoming diffs, when the class becomes polymorphic over the annotation position `'T`, it is a convenient way to enforce that this class is indeed polymorphic over `'T`. Otherwise, if some operation tried to unify `'T` with a concrete type like `ALoc.t * Type.t`, the type checker would not complain but the class would have a constraint inferred that would encode this unification.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55053041

fbshipit-source-id: 8d3731b92dc8f75503d24421e21dd2bfd81f7d7f