[flow][autocomplete] check enclosing statement when autocompleting error suppression in comments
Summary:
Autocomplete offers a special feature of returning an error suppression with the appropriate code, when invoked within a comment attached to a node that raises an error. To get this functionality in the on-demand mode of autocomplete we need to do some additional work, given that the existing context might not have the relevant error populated. To address this, when we are about to create an `Ac_comment` we also check the closest enclosing statement `stmt`, and run post-inference checks on `stmt`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55053042

fbshipit-source-id: 3cad4a7423474f544c3faf2e3b232a338f985124