[flow][refactor] Avoid using unresolved tvar for this param typing
Summary:
Right now, we still use unresolved tvar for this type of functions and methods, and this is an issue for LTI 2.0. I think there is no reason to do this anymore, since it's more of a leftover of LTI cleanup.

In a lot of places, we check whether `this` appears in function body or params. If so, we generate tvar which that ends up being resolved to any and error. If not, we use mixed as the type of this type. This is almost the full behavior of the this typing behavior, except that it misses the case that none of this matters if we do have `this` annotation.

In this diff, we add two functions that will allow us to complete the analysis locally: a function in signature_util that tells us whether we are missing this annotation, by additionally checking if the function already has this annotation, and a function in `flow_js_utils` that checks it and error accordingly. Then we use these functions to generate the types of default_this, instead of doing it in `mk_func_sig`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55171035

fbshipit-source-id: a713562fb0866e8f30b21323de8088d04d36683b