[flow] call post-inference checks within Type_inference_js.infer_file
Summary:
The calls to
* `Type_inference_js.infer_file`
* `Merge_js.post_merge_checks`
* `Flow_error.post_process_errors`

are always made together (except for some tests). It makes more sense to me to have these be called in the same function (`infer_ast` and `infer_lib_file`) instead of expecting the caller of `infer_file` to also call post-merge checking etc.

This diff also cleans up the API of `initialize_env` since some of the params passed in are already computable through information in `cx`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55144519

fbshipit-source-id: 4fd2828acf9bad02bd719c5e33ffa79d82197ee4