[flow] on-demand autocomplete
Summary:
This diff adds a mode to autocomplete that will infer types for the AST node where the request occurs. The mode is enabled through the option
```
experimental.autocomplete_mode=on_demand
```
In contrast with current mode (`typed_ast`), on-demand only needs to compute the typed environment for the content where the command runs. (See changes in command_handler.ml.)

The implementation of this mode leverages refactorings that happened leading to this diff. Specifically autocomplete_js.ml can run whether we have a typed-AST available or not (see `available_ast` type).

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55080918

fbshipit-source-id: a53a371393caa70dca7d6116bec7dcd7d20a50b0