[flow][tests] Print kind if it's not OrdinaryNameLoc in name_def_test
Summary:
Sometimes two kind_and_loc have the same locs. That's why we introduced kind in the first place. When two kind_and_loc that have the same loc appear in a test, it's confusing and doesn't help to enforce orders between things like function this vs function.

This diff prints the kind as well if it's not OrdinaryNameLoc. We skip the kind for OrdinaryNameLoc since it's the most common one.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55295037

fbshipit-source-id: 56276d9436380757943527fc44afb748e363fa00