[flow] Specialized implementation for identifier type synthesis in class extends
Summary:
The extends clause in class, although being an expression, is more annotation-like. We made this decision during LTI development so that we can tolerate more cyclic class definitions in the most common cases.

In the LTI 2.0 system, all the annotations will be resolved lazily all at once at the end of component resolution. Early forcing is not allowed, otherwise it will result in internal error. Currently, class extends' identifier are evaluated eagarly due to repositioning. We can achieve the same using AnnotT, and avoid the eagar repositioning. Using AnnotT to achieve repositioning effect is already used extensively in type_annotation and type_sig_merge when we are refering to a type identifier.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55296799

fbshipit-source-id: d9d035750e81d58eca4befdf1f947719779187f0