[flow][rfc] Builtin defense against cyclic tvars
Summary:
This diff implements a builtin defense against cyclic tvars for fully resolved lazy tvars, to ensure that tvar looping back to itself and then crashing is impossible by construction.

Unsurprisingly, due to our general lack of defense against cyclic types, and also it's impossible to rule them out without making some useful types un-expressible, we end up with cyclic tvars sometimes. If the tvar is unresolved, things might not be on fire, since the worst can happen is some unexpected empty, but if a lazy fully resolved tvar is cyclic, and when we force the lazy tvar, we will end up with `CamlinternalLazy.Undefined` crash, since there is just no way OCaml's lazy runtime can reasonable handle the case when we are forced to force a lazy value v when we are already in the middle of forcing the same value.

The defense built in this diff is designed to intercept right before the crash is about to happen, and return an any type with a potential side effect that adds an error. Intuitively, the lazy structure in OCaml has 3 state: unforced, in the middle of forcing, and forced. The bad condition for crashing is forcing while in the middle of forcing. In this diff, we also track the same 3 state ourselves, and perform appropriate logic when we force the lazy tvar in one of the 3 state.

Why we do need this? For LTI 2.0, we will make more heavy uses of fully resolved lazy tvars, to make the system more like annotation inference. We will need a principled way to defend it, similar to [how we defend it during annotation inference](https://github.com/facebook/flow/blob/518a3281d32a409b5526aeef4f03be98706189de/src/typing/annotation_inference.ml#L416-L418).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D54021755

fbshipit-source-id: 586a1ec936b9cc8abd8ac9e17bf725e44b84c879