[flow] do not allow recursion in %checks functions
Summary:
Functions like
```
function is_zero(x: number): boolean %checks {
  return is_zero(x);
}
```
even though they have a return annotation (`boolean`), will be considered unannotated for the purposes of determining whether a type can be inferred for them. Since the inferred predicate depends on the function body, and the body calls the function recursively, this will be a [recursive-definition] error.

The error message "Please add an annotation to this definition" implies that this is fixable through annotation, which is not the case. This is a minor point since %checks is deprecated, so fixing is left for future work.

Changelog: [errors] Predicate functions (`%checks`) cannot be recursive

Reviewed By: SamChou19815

Differential Revision: D53596999

fbshipit-source-id: 0ad17aa37475f7c1c31668f83dc35413508d9acb