[flow][name-def] Do not add FunctionThisLoc to resolve in classes for annotated fields
Summary:
When the class fields are annotated, this type of functions assigned to the fields are irrelevant to the class type. Therefore, we do not need to add them to be resolved together with classes.

Having this redundant step is problematic for LTI 2.0, which will require all the entries to be actually resolved at the end of component resolution.

During the process, I also find another bug that we are not handling get private prop and method for any receivers, and the change happens to uncover this. I added relevant constraints handling in flow_js to handle this case as well.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55295038

fbshipit-source-id: 2e675400c7fc016a818d77a4907cdb985095779b