[flow] More precise dependency analysis in class extends
Summary:
In the AST, class extends contain an expression, but during type checking, we do not accept any kind of expressions. In fact, we only accept those that are annotation-like, e.g. identifier, members, and casts. For casts, only the anntoation is used to determine the signature, and for all other expressions, the signature will be any.

However, our dependency analysis is name_def_ordering is not aware of these rules. We just blindly add everything in the expression as dependency. In the added test, it makes Flow think that there is a legal SCC of

```
  Ref -->  referencedInClassExtends --> Node --
   /\                                         |
    |-----------------------------------------\/
```

which doesn't make sense. It just has not blown up under the current system, due to the use of unresolved tvars in between annotation resolution.

This diff fixes that. During the dependency analysis, we will only visit the nodes that are used to determine the signature, mirroring the logic in `mk_extends` in statement.ml.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55528626

fbshipit-source-id: 62b7617b596395b133a8d9aea2f5395dc36aa0e8