[flow][refactor] Separate out export_search_types from export_search
Summary:
The goal is to make autocomplete work in try-flow.

Auto imports search, which depends on some C++ code, absolutely won't work in browser, so in this diff we did some work to separate that out. We split export search into the search parts and the types of the results. Later, we can make autocompleteService with configuration auto-import behavior only depend on export_search_types

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55548909

fbshipit-source-id: c4f1f950ef5c06d30f47ea6d5442b321876e577d