[flow][refactor] Remove autocompleteService's dependence on Options.t
Summary:
The goal is to make autocomplete work in try-flow. While this is not strictly a blocker, it's quite annoying to construct an `Options.t` in flow_dot_js when most of the stuff is used.

AutocompleteService currently depends on lsp_import_edits, which requires `Options.t`. All we need are a few options, so let's pass them directly.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55548907

fbshipit-source-id: 36431cc6c285247dbdb65146bda4bfca09c3b9ce