[flow][refactor] Split lsp into parts that are are pure and the parts that uses jsonrpc
Summary:
The goal is to make autocomplete work in try-flow.

Somewhere in autocomplete and code action, we will pull in the lsp target. This is understandable, since we want to produce these lsp response objects. However, the lsp target also includes some code that does some JSON-RPC stuff, which will not work in the browser.

This diff separates these two parts, so that when autocomplete and code actions pull in lsp, we only pull in these type definition and pure functions.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55549369

fbshipit-source-id: 3839e47e95790110de158e1605a56f1e6a429c12