[flow] Reduce the range of min/max safe integer in fuzzy_score
Summary:
The goal is to make autocomplete work in try-flow. This module will be pulled in to autocomplete, since it uses fuzzy score to filter and rank results.

For some reason, js_of_ocaml wants the integer to fit into 32 bits, so I reduced the range. It shouldn't really matter here, since the fuzzy score should never get that big/small.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55552711

fbshipit-source-id: 93a8f6f230b38bf06446c786b8d7c0222443c620