[flow][refactor] Remove autocompleteService's dependence on ServerProt
Summary:
The goal is to make autocomplete work in try-flow. ServerProt depends on logging, which pulls in some unix stuff, which won't work on flow.js.

Autocomplete service doesn't have to depend on it. We can just expose the AcComplete types and convert it into suitable format in other places.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55555440

fbshipit-source-id: 5b4eb361f3ab0f89bb43eaac258926306772d7a6