Preserve Opaque Types when evaluating ReactDRO destructor
Summary:
The ReactDRO destructor adds a flag to all object types that make that object deeply-read-only for local analysis. This is unique among destructors in that it isn't really *transforming* the type as it is *marking* the type.

Typically, we handle OpaqueTs in destructors by using their underlying type or otherwise falling back on behavior in flow_js.ml's larger pattern match when the underlying type is not exposable. For destructors that transform this makes sense: there's no sense in keeping around an opaque type because the opaque type is undergoing a transformation that would make it fundamentally different than its original definition. There's no sense in making that type opaque.

But for ReactDRO, we don't actually want to lose the opaque type, we just want to mark the underlying_t and super_t with the ReactDRO marker. It would make sense to preserve the opacity, and a DRO OpaqueT should be compatible with its non-DRO equivalent, since ReactDRO is just for local analysis.

This diff solves the issue by preserving opacity when evaluating ReactDRO on an opaque type. We have to go one extra step of forcing the evaluation of ReactDRO applied to the underlying_t and super_t because our refinement machinery expects to be able to inspect the underlying_t and super_t of opaque types.

Without the change to eagerly evaluate we get ~50 errors matching the conditional.js test case added.

Changelog: [fix] Fixed a behavior where component syntax would transform opaque types into their underlying types.

Reviewed By: SamChou19815

Differential Revision: D55575945

fbshipit-source-id: 64a6ca0febf717f6aa4e047e20a8a052f520abf1