Do not use hook types from libdefs when component syntax is not enabled
Summary:
We want to be able to declare hooks in the libdefs without opting people into react rule enforcement. THis is already mostly handled except for in type_sig_parse, where we will unconditionally produce artifacts that create hooks in type_sig_merge.

This diff fixes that issue by explicitly tracking if component syntax is enabled in the config. If it is not then we don't store hook-ness in signatures, even when hook syntax is enabled for the exporting file.

This enables us to enable hook libdefs, which is done in the next diff in the stack

Changelog: [internal]

Reviewed By: mvitousek

Differential Revision: D55624720

fbshipit-source-id: 6b82ba506c592b446488b5d2eeac9377fae28ba2