[flow][refactor] Remove signature help's dependence on ServerProt
Summary:
The goal is to make autocomplete work in try-flow. ServerProt depends on logging, which pulls in some unix stuff, which won't work on flow.js.

Signature help doesn't have to depend on it. We can just duplicate the type and convert it into suitable format in other places.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55555436

fbshipit-source-id: c4b71eb0ddc503678c13fc6f231ad18a26a36a13