[flow] Prevent method-unbinding errors when there is a mixed/any hint
Summary:
method-unbinding prevents errors by disallowing you from unbinding a method from the `this` required to call it. However, this behavior is incompatible with things like jest.expect:
```
// ERROR, cannot unbind method from foo
expect(foo.method).toBeCalled...
```

We can take advantage of LTI to make this experience significantly better. In this diff, when we see a hint of `mixed` (which cannot be called) or `any` (which indicates an explicit opt-out of safety) we allow a method to be unbound without any errors.

Reviewed By: mvitousek

Differential Revision: D54157530

fbshipit-source-id: 50f32fff815ff6582ed1828a4db5cd91eae21943