[flow-dev-tools] Do not remove suppressions that include "eslint-disable"
Summary:
While not common it is possible that a single comment is used to suppress both a Flow error and an eslint issue. E.g.
```
// $FlowFixMe eslint-disable-next-line no-fallthrough
```
This diff prevents `flow update-suppressions` from removing a Flow suppression comment if in included in the text associated with it is the string `eslint-disable`.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D55656739

fbshipit-source-id: 9768276f4df1895377347883dfef9a976f172815