[flow] Make the defense against cyclic tvars more robust
Summary:
It turns out that it's still possible to crash with cyclic tvars under the builtin defense.

In `ForcingState.map`, we are directly mapping a potentially bad cyclic lazy payload.

This diff adds in extra defense. I added a test (`invalid_self_recursive_mapped`) which will fail without the defense.

Note that these are edge cases. In practice, the lazy cyclic tvar we created are not that crazy and usually has extra indirections like `AnnotT` and `EvalT`, so that's why it hasn't blown up LTI 2.0 yet.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D55728897

fbshipit-source-id: ea3809edfbf47385ef125eb43d46f2561808a642